// src/app/features/customers/models/customer.ts
export interface Customer {

    id: number;

    customerCode: string;

    firstName: string;

    lastName: string;

    fullName: string;

    email: string;

    phone: string;

    address: string;

    city: string;

    state: string;

    country: string;

    postalCode: string;

    isActive: boolean;

}